mysite/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
Django==4.2.14
django-bootstrap-v5==1.0.11
django-filter==24.2
djangorestframework==3.15.2
Markdown==3.6
psycopg2-binary==2.9.9
soupsieve==2.5
sqlparse==0.5.0
